refactor(lotto): 패키지 분리
