+++ src/main/java/lotto/LottoController.java
@@ -1,0 +1,6 @@
+package lotto;
+
+public class LottoController {
+
+
+}
