+++ src/main/java/lotto/console/Constants.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.console;
 
 public final class Constants {
