+++ src/main/java/lotto/console/Input.java
@@ -1,0 +1,5 @@
+package lotto.console;
+
+public class Input {
+
+}
