+++ src/main/java/lotto/console/Output.java
@@ -1,0 +1,5 @@
+package lotto.console;
+
+public class Output {
+
+}
