+++ src/main/java/lotto/statistics/Statistics.java
@@ -1,8 +1,10 @@
-package lotto;
+package lotto.statistics;
 
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
+import lotto.ticket.Winning;
+import lotto.console.Constants;
 
 public class Statistics {
