+++ src/test/java/lotto/ApplicationTest.java
@@ -1,15 +1,15 @@
 package lotto;
 
-import camp.nextstep.edu.missionutils.test.NsTest;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import camp.nextstep.edu.missionutils.test.NsTest;
+import java.util.List;
+import org.junit.jupiter.api.Test;
+
 class ApplicationTest extends NsTest {
+
     private static final String ERROR_MESSAGE = "[ERROR]";
 
     @Test
@@ -52,6 +52,25 @@ class ApplicationTest extends NsTest {
             runException("1000j");
             assertThat(output()).contains(ERROR_MESSAGE);
         });
+    }
+
+    @Test
+    void 일등_테스트() {
+        assertRandomUniqueNumbersInRangeTest(
+                () -> {
+                    run("1000", "6,7,9,11,17,18", "45");
+                    assertThat(output()).contains(
+                            "1개를 구매했습니다.",
+                            "[6, 7, 9, 11, 17, 18]",
+                            "3개 일치 (5,000원) - 0개",
+                            "4개 일치 (50,000원) - 0개",
+                            "5개 일치 (1,500,000원) - 0개",
+                            "5개 일치, 보너스 볼 일치 (30,000,000원) - 0개",
+                            "6개 일치 (2,000,000,000원) - 1개"
+                    );
+                },
+                List.of(6, 7, 9, 11, 17, 18)
+        );
     }
 
     @Override
